[refactor](Nereids) bind slot, function and do type coercion in one rule
