+++ fe/fe-core/src/main/java/org/apache/doris/nereids/analyzer/NereidsAnalyzer.java
@@ -23,7 +23,6 @@ import org.apache.doris.nereids.jobs.batch.AnalyzeRulesJob;
 import org.apache.doris.nereids.jobs.batch.AnalyzeSubqueryRulesJob;
 import org.apache.doris.nereids.jobs.batch.CheckAnalysisJob;
 import org.apache.doris.nereids.jobs.batch.TypeCoercionJob;
-import org.apache.doris.nereids.rules.analysis.Scope;
 
 import java.util.Objects;
 import java.util.Optional;
@@ -33,6 +32,7 @@ import java.util.Optional;
  * TODO: revisit the interface after subquery analysis is supported.
  */
 public class NereidsAnalyzer {
+
     private final CascadesContext cascadesContext;
     private final Optional<Scope> outerScope;
