+++ fe/fe-core/src/main/java/org/apache/doris/nereids/pattern/MatchingContext.java
@@ -19,7 +19,7 @@ package org.apache.doris.nereids.pattern;
 
 import org.apache.doris.nereids.CascadesContext;
 import org.apache.doris.nereids.StatementContext;
-import org.apache.doris.nereids.rules.analysis.CTEContext;
+import org.apache.doris.nereids.analyzer.CTEContext;
 import org.apache.doris.nereids.trees.plans.Plan;
 import org.apache.doris.qe.ConnectContext;
