+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/analysis/Binder.java
@@ -1,0 +1,201 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package org.apache.doris.nereids.rules.analysis;
+
+import org.apache.doris.nereids.CascadesContext;
+import org.apache.doris.nereids.analyzer.Scope;
+import org.apache.doris.nereids.analyzer.UnboundAlias;
+import org.apache.doris.nereids.analyzer.UnboundSlot;
+import org.apache.doris.nereids.analyzer.UnboundStar;
+import org.apache.doris.nereids.exceptions.AnalysisException;
+import org.apache.doris.nereids.trees.expressions.Alias;
+import org.apache.doris.nereids.trees.expressions.BoundStar;
+import org.apache.doris.nereids.trees.expressions.Expression;
+import org.apache.doris.nereids.trees.expressions.NamedExpression;
+import org.apache.doris.nereids.trees.expressions.Slot;
+import org.apache.doris.nereids.trees.expressions.SlotReference;
+import org.apache.doris.planner.PlannerContext;
+
+import org.apache.commons.lang.StringUtils;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
+class Binder extends SubExprAnalyzer {
+
+    public Binder(Scope scope, CascadesContext cascadesContext) {
+        super(scope, cascadesContext);
+    }
+
+    public Expression bind(Expression expression) {
+        return expression.accept(this, null);
+    }
+
+    @Override
+    public Expression visitUnboundAlias(UnboundAlias unboundAlias, PlannerContext context) {
+        Expression child = unboundAlias.child().accept(this, context);
+        if (unboundAlias.getAlias().isPresent()) {
+            return new Alias(child, unboundAlias.getAlias().get());
+        }
+        if (child instanceof NamedExpression) {
+            return new Alias(child, ((NamedExpression) child).getName());
+        } else {
+            // TODO: resolve aliases
+            return new Alias(child, child.toSql());
+        }
+    }
+
+    @Override
+    public Slot visitUnboundSlot(UnboundSlot unboundSlot, PlannerContext context) {
+        Optional<List<Slot>> boundedOpt = Optional.of(bindSlot(unboundSlot, getScope().getSlots()));
+        boolean foundInThisScope = !boundedOpt.get().isEmpty();
+        // Currently only looking for symbols on the previous level.
+        if (!foundInThisScope && getScope().getOuterScope().isPresent()) {
+            boundedOpt = Optional.of(bindSlot(unboundSlot,
+                    getScope()
+                            .getOuterScope()
+                            .get()
+                            .getSlots()));
+        }
+        List<Slot> bounded = boundedOpt.get();
+        switch (bounded.size()) {
+            case 0:
+                // just return, give a chance to bind on another slot.
+                // if unbound finally, check will throw exception
+                return unboundSlot;
+            case 1:
+                if (!foundInThisScope) {
+                    getScope().getOuterScope().get().getCorrelatedSlots().add(bounded.get(0));
+                }
+                return bounded.get(0);
+            default:
+                // select t1.k k, t2.k
+                // from t1 join t2 order by k
+                //
+                // 't1.k k' is denoted by alias_k, its full name is 'k'
+                // 'order by k' is denoted as order_k, it full name is 'k'
+                // 't2.k' in select list, its full name is 't2.k'
+                //
+                // order_k can be bound on alias_k and t2.k
+                // alias_k is exactly matched, since its full name is exactly match full name of order_k
+                // t2.k is not exactly matched, since t2.k's full name is larger than order_k
+                List<Slot> exactMatch = bounded.stream()
+                        .filter(bound -> unboundSlot.getNameParts().size() == bound.getQualifier().size() + 1)
+                        .collect(Collectors.toList());
+                if (exactMatch.size() == 1) {
+                    return exactMatch.get(0);
+                }
+                throw new AnalysisException(String.format("%s is ambiguous: %s.",
+                        unboundSlot.toSql(),
+                        bounded.stream()
+                                .map(Slot::toString)
+                                .collect(Collectors.joining(", "))));
+        }
+    }
+
+    @Override
+    public Expression visitUnboundStar(UnboundStar unboundStar, PlannerContext context) {
+        List<String> qualifier = unboundStar.getQualifier();
+        List<Slot> slots = getScope().getSlots()
+                .stream()
+                .filter(slot -> !(slot instanceof SlotReference) || ((SlotReference) slot).isVisible())
+                .collect(Collectors.toList());
+        switch (qualifier.size()) {
+            case 0: // select *
+                return new BoundStar(slots);
+            case 1: // select table.*
+            case 2: // select db.table.*
+                return bindQualifiedStar(qualifier, slots);
+            default:
+                throw new AnalysisException("Not supported qualifier: "
+                        + StringUtils.join(qualifier, "."));
+        }
+    }
+
+    private BoundStar bindQualifiedStar(List<String> qualifierStar, List<Slot> boundSlots) {
+        // FIXME: compatible with previous behavior:
+        // https://github.com/apache/doris/pull/10415/files/3fe9cb0c3f805ab3a9678033b281b16ad93ec60a#r910239452
+        List<Slot> slots = boundSlots.stream().filter(boundSlot -> {
+            switch (qualifierStar.size()) {
+                // table.*
+                case 1:
+                    List<String> boundSlotQualifier = boundSlot.getQualifier();
+                    switch (boundSlotQualifier.size()) {
+                        // bound slot is `column` and no qualified
+                        case 0:
+                            return false;
+                        case 1: // bound slot is `table`.`column`
+                            return qualifierStar.get(0).equalsIgnoreCase(boundSlotQualifier.get(0));
+                        case 2:// bound slot is `db`.`table`.`column`
+                            return qualifierStar.get(0).equalsIgnoreCase(boundSlotQualifier.get(1));
+                        default:
+                            throw new AnalysisException("Not supported qualifier: "
+                                    + StringUtils.join(qualifierStar, "."));
+                    }
+                case 2: // db.table.*
+                    boundSlotQualifier = boundSlot.getQualifier();
+                    switch (boundSlotQualifier.size()) {
+                        // bound slot is `column` and no qualified
+                        case 0:
+                        case 1: // bound slot is `table`.`column`
+                            return false;
+                        case 2:// bound slot is `db`.`table`.`column`
+                            return qualifierStar.get(0).equalsIgnoreCase(boundSlotQualifier.get(0))
+                                    && qualifierStar.get(1).equalsIgnoreCase(boundSlotQualifier.get(1));
+                        default:
+                            throw new AnalysisException("Not supported qualifier: "
+                                    + StringUtils.join(qualifierStar, ".") + ".*");
+                    }
+                default:
+                    throw new AnalysisException("Not supported name: "
+                            + StringUtils.join(qualifierStar, ".") + ".*");
+            }
+        }).collect(Collectors.toList());
+
+        return new BoundStar(slots);
+    }
+
+    private List<Slot> bindSlot(UnboundSlot unboundSlot, List<Slot> boundSlots) {
+        return boundSlots.stream().filter(boundSlot -> {
+            List<String> nameParts = unboundSlot.getNameParts();
+            int qualifierSize = boundSlot.getQualifier().size();
+            int namePartsSize = nameParts.size();
+            if (namePartsSize > qualifierSize + 1) {
+                return false;
+            }
+            if (namePartsSize == 1) {
+                return nameParts.get(0).equalsIgnoreCase(boundSlot.getName());
+            }
+            if (namePartsSize == 2) {
+                String qualifierDbName = boundSlot.getQualifier().get(qualifierSize - 1);
+                return qualifierDbName.equalsIgnoreCase(nameParts.get(0))
+                        && boundSlot.getName().equalsIgnoreCase(nameParts.get(1));
+            } else if (nameParts.size() == 3) {
+                String qualifierDbName = boundSlot.getQualifier().get(qualifierSize - 1);
+                String qualifierClusterName = boundSlot.getQualifier().get(qualifierSize - 2);
+                return qualifierClusterName.equalsIgnoreCase(nameParts.get(0))
+                        && qualifierDbName.equalsIgnoreCase(nameParts.get(1))
+                        && boundSlot.getName().equalsIgnoreCase(nameParts.get(2));
+            }
+            //TODO: handle name parts more than three.
+            throw new AnalysisException("Not supported name: "
+                    + StringUtils.join(nameParts, "."));
+        }).collect(Collectors.toList());
+    }
+}
