+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/algebra/Scan.java
@@ -18,7 +18,6 @@
 package org.apache.doris.nereids.trees.plans.algebra;
 
 import org.apache.doris.catalog.TableIf;
-import org.apache.doris.nereids.analyzer.Relation;
 
 /**
  * Common interface for logical/physical scan.
