+++ fe/fe-core/src/test/java/org/apache/doris/nereids/parser/NereidsParserTest.java
@@ -173,7 +173,7 @@ public class NereidsParserTest extends ParserTestBase {
         String innerJoin2 = "SELECT t1.a FROM t1 JOIN t2 ON t1.id = t2.id;";
         logicalPlan = nereidsParser.parseSingle(innerJoin2);
         logicalJoin = (LogicalJoin) logicalPlan.child(0);
-        Assertions.assertEquals(JoinType.CROSS_JOIN, logicalJoin.getJoinType());
+        Assertions.assertEquals(JoinType.INNER_JOIN, logicalJoin.getJoinType());
 
         String leftJoin1 = "SELECT t1.a FROM t1 LEFT JOIN t2 ON t1.id = t2.id;";
         logicalPlan = nereidsParser.parseSingle(leftJoin1);
