+++ regression-test/data/nereids_p0/sql_functions/string_functions/test_string_function_regexp.out
@@ -19,13 +19,13 @@ billie eillish
 a-b-c
 
 -- !sql --
-a <\\1> c
+a <b> c
 
 -- !sql --
 a-b c
 
 -- !sql --
-a <\\1> b
+a <b> b
 
 -- !sql --
 false	1	1989	1001	11011902	123.123	true	1989-03-21	1989-03-21T13:00	wangjuoo4	0.1	6.333	string12345	170141183460469231731687303715884105727
